refactor the API to have one file per handler
