+++ requirements/requirements_api.txt
@@ -16,7 +16,7 @@ pandas==2.2.2
 pydantic==1.10.13 # bump to v2 once odmantic supports it
 PyJWT==2.8.0
 pymongo==4.7.2
-pytest==8.2.1
+pytest==8.1.2
 pytest-aiohttp==1.0.5
 PyYAML==6.0.1
 requests==2.32.0
